oz/pyproject.toml
[tool.poetry]
name = "five"
version = "0.1.0"
description = ""
authors = ["qnwlqnwlxm <qnwlqnwlxm@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.32"
pymysql = "^1.1.1"
faker = "^26.2.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
